requirements.txt
click==6.7
Flask==1.1.2
itsdangerous==0.24
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
numpy~=1.18.5
opencv-python~=4.2.0.34
opencv-contrib-python

requests~=2.23.0
Pillow~=7.1.2
